SMC fixed
